refactor(auth): move JWK validators to keyvalidation.go with exported names
trustedKIDPattern, parseRSAPublicKeyFromJWK, validateRSAPublicExponent
relocated. Exported MatchesTrustedKIDPattern and ParseRSAPublicKeyFromJWK
for adapter consumption. trusted.go body still present (deleted in a
later task when TrustedKeyStore interface lands).

Refs #281.
